refactor(desktop): move the insights IPC surface into its own module (CH-012)
The slop gate judges every changed file, and ipc.ts is over the
file-size limit before this item touches it (1553 lines at base;
issue #90 owns the split). Adding the insights wrappers there put a
pre-existing violation in front of the judge. The insights payload
types and command wrappers now live in insightsIpc.ts, ipc.ts returns
to its base state, and the four insights consumers import the new
module. No wrapper or type changes shape.
